test commit

+++ Infrastruktura/Repositories/ZamestnanecRepository.cs
@@ -7,11 +7,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Infrastruktura.Models;
 
-
-
 namespace Infrastruktura.Repositories
 {
-   
     public class ZamestnanecRepository
     {
         private readonly ZamestnanciContext _context;
@@ -21,7 +18,6 @@ namespace Infrastruktura.Repositories
             _context = context;
         }
 
-        
         public async Task <IEnumerable<Zamestnanci>> GetZamestnanci()
         {
             return await _context.Zamestnanci.ToListAsync();
@@ -36,11 +32,40 @@ namespace Infrastruktura.Repositories
 
         public async Task Put(int id, Zamestnanci zamestnanci)
         {
-              _context.Entry(zamestnanci).State = EntityState.Modified;
+             
+             
+            var zamestnanecPred = from zam in _context.Zamestnanci
+                                where zam.ZamestnanecId == id
+                                select zam.IdPozicie;
+            var zamestnanecPre = await zamestnanecPred.FirstOrDefaultAsync();
+
+            
+            
                 
+              _context.Entry(zamestnanci).State = EntityState.Modified;
+              
+              
             try
              {
                  await _context.SaveChangesAsync();
+                    // PredoslePozicie zmena = new PredoslePozicie();
+                    // zmena.ZamestnanecId=zamestnanec.ZamestnanecId;
+                    // zmena.PoziciaId=zamestnanec.IdPozicie;
+                    // zmena.DatumNastupu=zamestnanec.DatumNastupu;
+                    // zmena.DatumUkoncenia=DateTime.Now;
+                    //  _context.Predoslepozicie.Add(zmena);
+                    
+                    var zamestnanecPo = from pozmene in _context.Zamestnanci
+                                        where pozmene.ZamestnanecId == id
+                                        select pozmene.IdPozicie;
+
+                    if(zamestnanecPred != zamestnanecPo)
+                    { 
+                        _context.Predoslepozicie.Add(new PredoslePozicie{ZamestnanecId = zamestnanci.ZamestnanecId, PoziciaId = zamestnanci.IdPozicie, DatumNastupu = zamestnanci.DatumNastupu, DatumUkoncenia = DateTime.Now});
+                         await _context.SaveChangesAsync();
+                    }
+
+                 
              }
              catch (DbUpdateConcurrencyException)
              {
@@ -53,7 +78,7 @@ namespace Infrastruktura.Repositories
 
             private bool ZamestnanciExists(int id)
             {
-                return _context.Zamestnanci.Any(e => e.Id == id);
+                return _context.Zamestnanci.Any(e => e.ZamestnanecId == id);
             }
        
         public async Task PostZamestnanci(Zamestnanci zamestnanci)
@@ -62,11 +87,21 @@ namespace Infrastruktura.Repositories
              await _context.SaveChangesAsync();
         }
 
-         public async Task DeleteZamestnanci(int id)
+         public async Task DeleteZamestnanci(int id, bool archivovat)
          {
+
              var zamestnanci = await _context.Zamestnanci.FindAsync(id);
-             _context.Zamestnanci.Remove(zamestnanci);
-             await _context.SaveChangesAsync();
+
+                if (archivovat)
+                {
+                    zamestnanci.Archivovany = true;
+                }
+                else 
+                {
+                    _context.Zamestnanci.Remove(zamestnanci);
+                }
+
+                    await _context.SaveChangesAsync();
          }
